requirements.txt
# conda create -n rhde python=3.8.11
# conda activate rhde
# pip install -r requirements.txt -f https://download.pytorch.org/whl/torch_stable.html
numpy==1.23.0
scipy==1.5.2
six==1.16.0
Pillow==10.4.0
opencv-python==4.10.0.84
torch==1.7.1+cu110
torchvision==0.8.2+cu110

# # pip install dlib-19.21.1-cp38-cp38-linux_x86_64.whl
# facenet_pytorch==2.4.1
# # git clone https://github.com/YadiraF/face3d
# # cd face3d/mesh/cython
# # python setup.py build_ext -i 
# scikit-image==0.17.2
# matplotlib==3.7.5
# Cython==3.0.11